+++ .github/workflows/SearchAdCreatives_build_and_test_on_main.yml
@@ -13,6 +13,8 @@ jobs:
     uses: FrendsPlatform/FrendsTasks/.github/workflows/linux_build_main.yml@main
     with:
       workdir: Frends.LinkedIn.SearchAdCreatives
+      env_var_name_1: Frends_LinkedIn_AccessToken
     secrets:
       badge_service_api_key: ${{ secrets.BADGE_SERVICE_API_KEY }}
+      env_var_value_1: ${{ secrets.FRENDS_LINKEDIN_ACCESSTOKEN }}
