+++ .github/workflows/SearchAdCreatives_build_and_test_on_push.yml
@@ -13,6 +13,8 @@ jobs:
     uses: FrendsPlatform/FrendsTasks/.github/workflows/linux_build_test.yml@main
     with:
       workdir: Frends.LinkedIn.SearchAdCreatives
+      env_var_name_1: Frends_LinkedIn_AccessToken
     secrets:
       badge_service_api_key: ${{ secrets.BADGE_SERVICE_API_KEY }}
       test_feed_api_key: ${{ secrets.TASKS_TEST_FEED_API_KEY }}
+      env_var_value_1: ${{ secrets.FRENDS_LINKEDIN_ACCESSTOKEN }}
