+++ Frends.LinkedIn.SearchAdCreatives/Frends.LinkedIn.SearchAdCreatives.Tests/Frends.LinkedIn.SearchAdCreatives.Tests.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFrameworks>net6.0;net471</TargetFrameworks>
 
     <IsPackable>false</IsPackable>
   </PropertyGroup>
