+++ Frends.LinkedIn.SearchAdCreatives/Frends.LinkedIn.SearchAdCreatives/Definitions/AdCampaign.cs
@@ -1,0 +1,13 @@
+﻿namespace Frends.LinkedIn.SearchAdCreatives.Definitions
+{
+    /// <summary>
+    /// Class for Campaigns.
+    /// </summary>
+    public class AdCampaign
+    {
+        /// <summary>
+        /// Campaign URN for the Campaign used as filter.
+        /// </summary>
+        public string AdCampaignUrn { get; set; }
+    }
+}
