+++ Frends.LinkedIn.SearchAdCreatives/Frends.LinkedIn.SearchAdCreatives/Definitions/AdCreative.cs
@@ -1,0 +1,13 @@
+﻿namespace Frends.LinkedIn.SearchAdCreatives.Definitions
+{
+    /// <summary>
+    /// Class for AdCreative.
+    /// </summary>
+    public class AdCreative
+    {
+        /// <summary>
+        /// AdCreative URN for the Caretives used as filter.
+        /// </summary>
+        public string AdCreativeUrn { get; set; }
+    }
+}
