+++ Frends.LinkedIn.SearchAdCreatives/Frends.LinkedIn.SearchAdCreatives/Frends.LinkedIn.SearchAdCreatives.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	  <TargetFramework>net6.0</TargetFramework>
+	  <TargetFrameworks>net6.0;netstandard2.0;net471</TargetFrameworks>
 	  <LangVersion>Latest</LangVersion>
 	  <Version>1.0.0</Version>
 	  <Authors>Frends</Authors>
@@ -15,6 +15,10 @@
 	  <PackageProjectUrl>https://frends.com/</PackageProjectUrl>
 	  <RepositoryUrl>https://github.com/FrendsPlatform/Frends.LinkedIn/tree/main/Frends.LinkedIn.SearchAdCreatives</RepositoryUrl>
   </PropertyGroup>
+
+	<ItemGroup Condition="'$(TargetFramework)'=='net471'">
+		<Reference Include="System.Net.Http" />
+	</ItemGroup>
 
   <ItemGroup>
     <None Include="FrendsTaskMetadata.json" Pack="true" PackagePath="/">
