+++ Frends.LinkedIn.SearchAdCreatives/Frends.LinkedIn.SearchAdCreatives/HttpClientFactory.cs
@@ -1,14 +1,15 @@
-﻿namespace Frends.LinkedIn.SearchAdCreatives;
-
-using System.Net.Http;
-using Frends.LinkedIn.SearchAdCreatives.Definitions;
-
-internal class HttpClientFactory : IHttpClientFactory
+﻿namespace Frends.LinkedIn.SearchAdCreatives
 {
-    public HttpClient CreateClient(Options options)
+    using System.Net.Http;
+    using Frends.LinkedIn.SearchAdCreatives.Definitions;
+
+    internal class HttpClientFactory : IHttpClientFactory
     {
-        var handler = new HttpClientHandler();
-        handler.SetHandlerSettingsBasedOnOptions(options);
-        return new HttpClient(handler);
+        public HttpClient CreateClient(Options options)
+        {
+            var handler = new HttpClientHandler();
+            handler.SetHandlerSettingsBasedOnOptions(options);
+            return new HttpClient(handler);
+        }
     }
 }
